+++ DepthRenderer/__main__.py
@@ -41,6 +41,11 @@ def resize(image, size):
         kind='option',
         type=int
     ),
+    displacement_factor=plac.Annotation(
+        help='A multiplicative constant applied to the normalised depth values.',
+        kind='option',
+        type=float
+    ),
     output_path=plac.Annotation(
         help='The path to save any output frames to.',
         kind='option',
@@ -48,7 +53,7 @@ def resize(image, size):
     )
 )
 def main(image_path="samples/00000_colors.png", depth_path="samples/00000_depth.png",
-         fps=60, mesh_density=8, output_path='frames'):
+         fps=60, mesh_density=8, displacement_factor=4.0, output_path='frames'):
     """
     Render a colour/depth image pair on a grid mesh in OpenGL using the depth map to displace vertices on the mesh.
 
@@ -57,6 +62,8 @@ def main(image_path="samples/00000_colors.png", depth_path="samples/00000_depth.
     :param fps: The target frames per second at which to render.
     :param mesh_density: How fine the generated mesh should be. Increasing this value by one roughly quadruples the
         number of vertices.
+    :param displacement_factor: A multiplicative constant applied to the normalised depth values.
+        For example, with a displacement factor of 10.0, a normalised depth value of 0.2 becomes 10.0 * 0.2 = 2.0.
     :param output_path: The path to save any output frames to.
     """
     colour = load_colour(image_path)
@@ -65,9 +72,10 @@ def main(image_path="samples/00000_colors.png", depth_path="samples/00000_depth.
 
     texture = Texture(colour)
     mesh = Mesh.from_depth_map(texture, density=mesh_density, depth_map=depth)
+    mesh.vertices[:, 2] *= displacement_factor
 
-    camera_position = get_translation_matrix(dz=-100)
-    camera = Camera(window_size=tuple(reversed(colour.shape[:2])), fov_y=180)
+    camera_position = get_translation_matrix(dz=-10)
+    camera = Camera(window_size=tuple(reversed(colour.shape[:2])), fov_y=18)
 
     # TODO: Make shader source paths configurable.
     default_shader = ShaderProgram(vertex_shader_path='DepthRenderer/shaders/shader.vert',
@@ -75,11 +83,11 @@ def main(image_path="samples/00000_colors.png", depth_path="samples/00000_depth.
     debug_shader = ShaderProgram(vertex_shader_path='DepthRenderer/shaders/debug_shader.vert',
                                  fragment_shader_path='DepthRenderer/shaders/debug_shader.frag')
 
-    renderer = MeshRenderer(mesh,
-                            default_shader_program=default_shader,
+    renderer = MeshRenderer(default_shader_program=default_shader,
                             debug_shader_program=debug_shader,
                             fps=fps, camera=camera)
 
+    renderer.mesh = mesh
     mesh.transform = get_scale_matrix(1.0) @ mesh.transform
     log(f"Model: \n{mesh.transform}")
 
@@ -92,18 +100,19 @@ def main(image_path="samples/00000_colors.png", depth_path="samples/00000_depth.
 
     animation_speed = 0.15
 
-    anim = RotateXYBounce(np.deg2rad(2.5), offset=0.5, speed=-animation_speed)
+    anim = Compose([
+        RotateAxisBounce(np.deg2rad(2.5), axis=Axis.Y, offset=0.5, speed=-animation_speed),
+        RotateAxisBounce(np.deg2rad(.5), axis=Axis.X, offset=0.5, speed=-animation_speed),
+        Translate(distance=0.30, speed=animation_speed),
+        Translate(distance=0.15, axis=Axis.Y, offset=0.25, speed=animation_speed)
+    ])
 
-    writer = AsyncImageWriter(renderer.frame_buffer_shape,
-                              num_workers=1)
+    writer = AsyncImageWriter(num_workers=1)
 
     render_one_frame = DelayedTask(OneTimeTask(writer.write), delay=10)
 
     video_writer = AsyncVideoWriter(os.path.join(output_path, f"{Path(image_path).name}.avi"),
-                                    size=renderer.frame_buffer_shape,
-                                    fourcc=cv2.VideoWriter_fourcc(*'DIVX'),
-                                    fps=fps,
-                                    num_workers=1)
+                                    size=renderer.frame_buffer_shape, fourcc=cv2.VideoWriter_fourcc(*'DIVX'), fps=fps)
 
     # Need to delay writing by one frame to ensure the window size is settled (#TODO: Find out why it's changing one after init)
     initial_delay = 3
@@ -113,12 +122,13 @@ def main(image_path="samples/00000_colors.png", depth_path="samples/00000_depth.
     def update_callback(delta):
         anim.update(delta)
 
+        # TODO: Fix bug that ignores panning from mouse input. Need to store transforms to the view from mouse inputs
+        #  separately instead of in Camera.view so that they are not overridden here.
         camera.view = camera_position @ anim.transform
 
-        # TODO: Fix `[mpeg4 @ 000001a3c8fa7dc0] Invalid pts (1) <= last (1)` warnings/errors.
-        # TODO: Fix intermittent SIGSEV and other errors.
-        render_one_frame(renderer.frame_buffer, os.path.join(output_path, 'sample_frame.png'))
-        write_frame(renderer.frame_buffer)
+        frame = renderer.get_frame()
+        render_one_frame(frame, os.path.join(output_path, 'sample_frame.png'))
+        write_frame(frame)
 
     def on_exit_callback():
         video_writer.cleanup()